+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 2.8.5
+
+Adds support for progressively loading the contents of a collection view that can display a very large number of items. The following new properties can be used to control this behaviour:
+ - `AdditionalData` is an infotable that, when updated, is added to the rows of the `Data` property.
+ - `DataCurrentSize` is a binding source that can be used to determine how many items are currently loaded.
+ - `DataTotalSize` is a binding target that can be used to specify how many total items there will be when everything is loaded.
+ - `HasCompleteDataSet` is both a binding target and source that can be used to specify when collection view should stop requesting additional data and is automatically set to `true` when `DataCurrentSize` becomes equal to `DataTotalSize`.
+ - `DataSetEndThreshold` controls how early the `CollectionViewWillApproachDataSetEnd` event will be triggered by specifying the distance to the data set end at which it should trigger.
+ - `PreventsRepeatedDataEndEvents` is a property that is enabled by default. When enabled, the two data set end events will only fire once, until the `AdditionalData` property is updated.
+
+The following new events can be used to request additional data:
+ - `CollectionViewWillApproachDataSetEnd` is triggered when the scroll position approaches the final items in the current data set.
+ - `CollectionViewDidReachDataSetEnd` is triggered when the scroll position has reached the end of the current data set.
+
 # 2.8.2
 
 Resolves an issue that caused mashups displayed by collection views to have incorrect styling when they had spaces in their name.